Factor simd_op_check into separate files.
